+++ Windows/SettingsWindow.xaml.cs
@@ -4,6 +4,8 @@
     using System;
     using System.Windows;
     using System.Windows.Controls;
+    using System.Windows.Controls.Primitives;
+    using Properties;
 
     /// <summary>
     /// Interaction logic for SettingsWindow.xaml
@@ -14,6 +16,16 @@
 
         private void DiscordRPforVSSettingsWindow_Loaded(Object sender, RoutedEventArgs e)
         {
+            this.IsPresenceEnabled.Content = Translates.SettingsWindow.IsPresenceEnabled(Settings.Default.translates);
+            this.IsFileNameShown.Content = Translates.SettingsWindow.IsFileNameShown(Settings.Default.translates);
+            this.IsSolutionNameShown.Content = Translates.SettingsWindow.IsSolutionNameShown(Settings.Default.translates);
+            this.IsTimestampShown.Content = Translates.SettingsWindow.IsTimestampShown(Settings.Default.translates);
+            this.IsTimestampResetEnabled.Content = Translates.SettingsWindow.IsTimestampResetEnabled(Settings.Default.translates);
+            this.IsLanguageImageLarge.Content = Translates.SettingsWindow.IsLanguageImageLarge(Settings.Default.translates);
+            this.SecretMode.Content = Translates.SettingsWindow.SecretMode(Settings.Default.translates);
+            this.LoadOnStartup.Content = Translates.SettingsWindow.LoadOnStartup(Settings.Default.translates);
+            this.SaveButton.Content = Translates.SettingsWindow.SaveButton(Settings.Default.translates);
+
             this.IsPresenceEnabled.IsChecked = DiscordRPforVSPackage.Settings.enabled;
             this.IsFileNameShown.IsChecked = DiscordRPforVSPackage.Settings.showFileName;
             this.IsSolutionNameShown.IsChecked = DiscordRPforVSPackage.Settings.showSolutionName;